Added new EmptyTable component with loading indicators to positions table.
