+++ src/views/Market/components/EmptyTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EmptyTable'
